% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{base_distribution}
\alias{base_distribution}
\title{Create a graphical evaluation of sequencing results}
\usage{
base_distribution(input_sequence, ab1file, replacements,
  trace_cutoff = 80)
}
\arguments{
\item{input_sequence}{The sequence which was modified. This is an object of type character containing the sequence.}

\item{ab1file}{The path to the ab1file which was provided by the sequencer/sequencing service}

\item{replacements}{The mutations which were desired.}

\item{trace_cutoff}{The minimal sum of signals (4 nucleotides) for a position in the sequence. [default: 80]}
}
\value{
Plots on the active/default graphics device.
}
\description{
This function creates a graphical evalution of the sequencing results to determine the quality of the created library.
How it works: 
The functions aligns the given input_sequence to the sequenced sequence (it also tries to align the reverse complement). Afterwards it searches for mismatches between the sequences.
Mismatches can be sucessfully mutated nucleotides. For the positions with a mismatch a pie chart showing the distribution of signals for each nucleotide is created.
You can controll the quality of the created library by comparing the pie chart to your expections about the modidication of the sequence.
}
\examples{
\dontrun{
data(MSD_BsaI_setup_lv2)
abfile<-"activesite_for_200718.ab1"
base_distribution(input_sequence=input_sequence, ab1file=abfile, replacements=mutations)
}
}
