% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_tm_nnb}
\alias{calculate_tm_nnb}
\title{Calculate melting temperature based on next neighbor calculation}
\usage{
calculate_tm_nnb(
  oligo_sequence,
  primer_concentration = 50,
  salt_concentration = 50,
  offset = 0
)
}
\arguments{
\item{oligo_sequence}{A string containing an oligo sequence.}

\item{primer_concentration}{The concentration of the primer in nanomole [default: 50]}

\item{salt_concentration}{The concentration of Na+ in nanomole [default: 50]}

\item{offset}{You can skip a prefix of your oligo sequence with this parameter. The first n bases are not considered in the calculation. [default: 0]}
}
\value{
An array or a list with values for the codons/amino acids.

The melting temperature in \code{print('\u00B0')}C
}
\description{
The implementation is based on the explanations of \url{http://biotools.nubic.northwestern.edu/OligoCalc.html}.
}
\details{
More details at \url{https://doi.org/10.1093/nar/gkm234}
}
\examples{
\dontrun{
GoldenMutagenesis::calculate_tm_nnb("AAAAAATGGTGTGTGATGTGTCCCTCTATC")
}

}
