% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuf}
\alias{cuf}
\title{The Codon Usage Table which is being used to select the codon for an exchanged amino acid}
\description{
The Codon Usage Table which is being used to select the codon for an exchanged amino acid
}
\seealso{
[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]
}
