% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{domesticate}
\alias{domesticate}
\title{Domestication of the input sequence}
\usage{
domesticate(input_sequence, restriction_enzyme = "GGTCTC",
  cuf = "e_coli_316407.csv")
}
\arguments{
\item{input_sequence}{The sequence which should be modified. This is just an object of type character containing the sequence.}

\item{restriction_enzyme}{Recognition site sequence of the respective restriction enzyme [default: GGTCTC]}

\item{cuf}{The Codon Usage Table which is being used to select the codon for an exchanged amino acid (and in this case to select the codon which shoulb be replaced). [default: e_coli_316407.csv]}
}
\value{
A list with replacments: Each element has a vector with the codon number at the first slot and the amino acid of this position at the second slot.
}
\description{
The domesticate function checks for internal cleavage sites. If corresponding sites are present silent mutations are introduced to destroy the recognition sites. 
The functions returns a list containing the position of the choosen amino acid residue for silent mutation.
}
