% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cu_table}
\alias{get_cu_table}
\title{Select a Codon Usage Table}
\usage{
get_cu_table(name, list = T)
}
\arguments{
\item{name}{The filname of the codon usage table. You can use list_cu_table() to get the overview.}

\item{list}{A boolean parameter that decides if an array or a list is returned. The array output is requiered by the domesticate function}
}
\value{
An array or a list with values for the codons/amino acids.
}
\description{
Get a list or an array of values for the selected Codon Usage Table
}
\examples{
list_cu_table()
\dontrun{
get_cu_table("e_coli_316407.csv")
}

}
