% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_sequence}
\alias{input_sequence}
\title{The sequence which should be modified. 
This is an object of type character containing the sequence}
\description{
The sequence which should be modified. 
This is an object of type character containing the sequence
}
\seealso{
[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]
}
