% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{list_cu_table}
\alias{list_cu_table}
\title{List the available Codon Usage Tables}
\usage{
list_cu_table()
}
\value{
A list of files
}
\description{
List the available Codon Usage Tables
}
\examples{
list_cu_table()
}
