% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSD_BsaI_result_lv2}
\alias{MSD_BsaI_result_lv2}
\title{Designed primers for Multiple Site Directed Mutagenesis}
\format{Primer object which is the result of the calculation}
\source{
\url{https://github.com/ipb-halle/GoldenMutagenesis/tree/master/vignettes}
}
\description{
This is the result of the vignette for Multiple Site Directed Mutagenesis in Level 2 with BsaI
}
