% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSD_BsaI_setup_lv2}
\alias{MSD_BsaI_setup_lv2}
\title{Basic setup for Multiple Site Directed Mutagenesis}
\format{The following objects are included:
 \describe{
  \item{cuf}{The Codon Usage Table which is being used to select the codon for an exchanged amino acid.}
  \item{input_sequence}{The sequence which should be modified. This is just an object of type character containing the sequence. }
  \item{mutations}{The desired mutations.}
  \item{recognition_site_bsai}{The BsaI recognition site.}
  
}}
\source{
\url{https://github.com/ipb-halle/GoldenMutagenesis/tree/master/vignettes}
}
\description{
Those are the objects of the MSD vignette.
}
