% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{mutate}
\alias{mutate}
\title{Calculate primers for a Point Mutagenesis}
\usage{
mutate(input_sequence, prefix = "TT", restriction_enzyme = "GGTCTC",
  suffix = "A", vector = c("AATG", "AAGC"), replacements,
  binding_min_length = 4, primer_length = 9, target_temp = 60,
  cuf = "e_coli_316407.csv")
}
\arguments{
\item{input_sequence}{The sequence which should be modified. This is just an object of type character containing the sequence.}

\item{prefix}{Additional nucleobases in 5' position of the recognition site [default: TT]}

\item{restriction_enzyme}{Recognition site sequence of the respective restriction enzyme [default: GGTCTC]}

\item{suffix}{Spacer nucleotides matching the cleavage pattern of the enzyme [default: A]}

\item{vector}{Four basepair overhangs complementary to the created overhangs in the acceptor vector  [default: c("AATG", "AAGC")]}

\item{replacements}{The desired substitutions}

\item{binding_min_length}{The minimal threshold value of the template binding sequence [default: 4]}

\item{primer_length}{Maximal length of the binding sequence [default: 9]}

\item{target_temp}{Melting temperature of the binding sequence in °C [default: 60]}

\item{cuf}{The Codon Usage Table which is being used to select the codon for an exchanged amino acid. [default: e_coli_316407.csv]}
}
\value{
An object of class Primerset with the designed Primers.
}
\description{
The mutate function designs the necessary set of primers for the desired mutations.
An example is given in the vignette at \url{https://github.com/ipb-halle/GoldenMutagenesis/blob/master/vignettes/Point_Mutagenesis.md}
}
\examples{

}
