% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{mutate_spm}
\alias{mutate_spm}
\title{Calculate Primers for Point Mutagenesis}
\usage{
mutate_spm(input_sequence, prefix = "TT",
  restriction_enzyme = "GGTCTC", suffix = "A", vector = c("AATG",
  "AAGC"), replacements, replacement_range = 2, binding_min_length = 4,
  binding_max_length = 9, target_temp = 60,
  cuf = "e_coli_316407.csv", fragment_min_size = 100)
}
\arguments{
\item{input_sequence}{The sequence which should be modified. This is an object of type character containing the sequence.}

\item{prefix}{Additional nucleobases in 5' position of the recognition site [default: TT]}

\item{restriction_enzyme}{Recognition site sequence of the respective restriction enzyme [default: GGTCTC]}

\item{suffix}{Spacer nucleotides matching the cleavage pattern of the enzyme [default: A]}

\item{vector}{Four basepair overhangs complementary to the created overhangs in the acceptor vector  [default: c("AATG", "AAGC")]}

\item{replacements}{The desired substitutions}

\item{replacement_range}{Maximum distance in amino acid residues between two randomization sites to be incoporated into a single primer (reverse, end of the fragment) - has a cascading effect for following mutations [default: 2]}

\item{binding_min_length}{The minimal threshold value of the length of the template binding sequence in amino acid residues [default: 4]}

\item{binding_max_length}{Maximal length of the binding sequence in amino acid residues [default: 9]}

\item{target_temp}{Melting temperature of the binding sequence in \code{print('\u00B0')}C [default: 60]}

\item{cuf}{The Codon Usage Table which is being used to select the codon for an exchanged amino acid. [default: e_coli_316407.csv]}

\item{fragment_min_size}{Minimal size of a generated gene fragment in base pairs [default 100]}
}
\value{
An object of class Primerset with the designed Primers.
}
\description{
The mutate function designs the necessary set of primers for the desired mutations.
An example is given in the vignette at \url{https://github.com/ipb-halle/GoldenMutagenesis/blob/master/vignettes/Point_Mutagenesis.md}
}
\examples{
#Load the setup of the Point Mutation vignette and design the primers
data(Point_Mutagenesis_BbsI_setup)
primers<-mutate_spm(input_sequence, prefix="TT", restriction_enzyme = recognition_site_bbsi, 
suffix = "AA", vector=c("CTCA", "CTCG"), replacements = mutations, binding_min_length=4 ,
binding_max_length=9, target_temp=60, cuf=cuf)

}
