% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{order_replacements}
\alias{order_replacements}
\title{Order the replacement list}
\usage{
order_replacements(replacements)
}
\arguments{
\item{replacements}{A list containing vectors which have a number at the first slot and the amino acid at the second one}
}
\value{
A sorted list of the input
}
\description{
Order the replacement list
}
\examples{
\dontrun{
replacements<-list(c(55, "V"), c(40, "A"))
GoldenMutagenesis::order_replacements(replacements)
}
}
