% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Point_Mutagenesis_BbsI_result}
\alias{Point_Mutagenesis_BbsI_result}
\title{Designed primers for Point Mutagenesis}
\format{Primer object which is the result of the calculation}
\source{
\url{https://github.com/ipb-halle/GoldenMutagenesis/tree/master/vignettes}
}
\description{
This is the result of the vignette for Point Mutagenesis with BbsI
}
