% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{primer_add_level}
\alias{primer_add_level}
\title{Add a level to exisiting Primerset}
\usage{
primer_add_level(primerset, prefix = "TT",
  restriction_enzyme = "GAAGAC", suffix = "AA", vector = c("CTCA",
  "CTCG"))
}
\arguments{
\item{primerset}{An exisiting Primerset (in Level 2)}

\item{prefix}{Additional nucleobases in 5' position of the new recognition site [default: TT]}

\item{restriction_enzyme}{Recognition site sequence of the new restriction enzyme (Level 0) [default: GAAGAC]}

\item{suffix}{Spacer nucleotides matching the cleavage pattern of the enzyme (Level 0) [default: AA]}

\item{vector}{Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("CTCA", "CTCG")]}
}
\value{
A Primerset in the new Level (Level 0)
}
\description{
This function replaces the prefix, the suffix and the restriction enzyme of a given Primerset to change the design to another Level.
You can use this function to convert an exisiting Level 2 Primerset to a Level 0 Primerset for example.
Also the overhangs of the first and the last primer will be modified to match the plasmid of the new level.
}
\examples{
#Load level 2 results of the MSD vignette
data(MSD_BsaI_result_lv2)
primer_add_level(primers,  prefix="TT", 
restriction_enzyme="GAAGAC", suffix="AA", vector=c("CTCA", "CTCG"))

}
