% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{primer_prepare_level}
\alias{primer_prepare_level}
\title{Prepare a Single Point Mutation Primerset to be used in Level 2}
\usage{
primer_prepare_level(primerset, vector = c("AATG", "AAGC"))
}
\arguments{
\item{primerset}{An exisiting Primerset (in Level 0)}

\item{vector}{Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("A", "AAGC")]}
}
\value{
A Primerset prepared for expression in Level 2
}
\description{
This function adds definied vector overhangs to Level 0 Primersets to express them in a Level 2 vector.
#'
}
\examples{
#Load level 0 results of the SPM vignette
data(Point_Mutagenesis_BbsI_result)
primer_prepare_level(primers)

}
