% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primers}
\alias{primers}
\title{Generated Primers
From Multiple Site Directed Mutagenesis or
Single Point Mutation}
\description{
Generated Primers
From Multiple Site Directed Mutagenesis or
Single Point Mutation
}
\seealso{
[Point_Mutagenesis_BbsI_result] and [MSD_BsaI_result_lv2]
}
