% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\docType{methods}
\name{print_primer}
\alias{print_primer}
\alias{print_primer,Primer-method}
\alias{print_primer,Primer_MSD-method}
\alias{print_primer,Primerset-method}
\alias{print_primer,Extended_Primerset-method}
\title{A function to generate a slim ouput text for primers}
\usage{
print_primer(primer)

\S4method{print_primer}{Primer}(primer)

\S4method{print_primer}{Primer_MSD}(primer)

\S4method{print_primer}{Primerset}(primer)

\S4method{print_primer}{Extended_Primerset}(primer)
}
\arguments{
\item{primer}{An object of class Primer, Primer_MSD or Primerset}
}
\value{
Textual output
}
\description{
An example is shown in the vignette at \url{https://github.com/ipb-halle/GoldenMutagenesis/blob/master/vignettes/Point_Mutagenesis.md}
}
\examples{
#Load results of the Point Mutation vignette and print it
data(Point_Mutagenesis_BbsI_result)
print_primer(primers)
}
