% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recognition_site_bbsi}
\alias{recognition_site_bbsi}
\title{The BbsI recognition site}
\description{
The BbsI recognition site

The BbsI recognition site.

The BbsI recognition site.
}
\seealso{
[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]

[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]

[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]
}
