% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recognition_site_bsai}
\alias{recognition_site_bsai}
\title{The BsaI recognition site}
\description{
The BsaI recognition site
}
\seealso{
[Point_Mutagenesis_BbsI_setup] and [MSD_BsaI_setup_lv2]
}
