% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{base_distribution}
\alias{base_distribution}
\title{Create a graphical evaluation of sequencing results}
\usage{
base_distribution(input_sequence, ab1file, replacements,
  trace_cutoff = 80)
}
\arguments{
\item{input_sequence}{The sequence which was modified. This is an object of type character containing the sequence.}

\item{ab1file}{The path to the ab1file which was provided by the sequencer/sequencing service}

\item{replacements}{The mutations which were desired.}

\item{trace_cutoff}{The minimal sum of signals (4 nucleotides) for a position in the sequence. [default: 80]}
}
\value{
Plots on the active/default graphics device.
}
\description{
This function creates a graphical evalution of the sequencing results to determine the quality of the created library.
}
\details{
The functions aligns the obtained sequencing results to the target gene sequence. 
It also tries to align the reverse complement of the obtained sequence. 
Afterwards it checks for mismatches between the sequences.
Mismatches are likely to be sucessfully mutated nucleotides. 
Positions regarded as mismatches are displayed as pie charts. 
The shown distributions are based on the signal intensities of the four nucleobases at the mismatch positions.
You can compare the pie charts with expected pattern of randomization, therefore validating the quality of the created library.
}
\examples{
\dontrun{
data(MSD_BsaI_setup_lv2)
abfile<-"activesite_for_200718.ab1"
base_distribution(input_sequence=input_sequence, ab1file=abfile, replacements=mutations)
}
}
