% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{print_primer}
\alias{print_primer}
\title{A function to generate a slim ouput text for primers}
\usage{
print_primer(primer)
}
\arguments{
\item{primer}{An object of class Primer, Primer MSD or Primerset}
}
\value{
A textual output
}
\description{
An example is shown in the vignette at \url{https://github.com/ipb-halle/GoldenMutagenesis/blob/master/vignettes/Point_Mutagenesis.md}
}
\examples{

}
